+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 - Update to JUnit 6.0.2
 - Retire Apache Derby integration (#2866)
 - Run test suite against Spring Framework 7 (#2919)
+- Add OraclePlugin that sets untyped null argument to Types.NULL for Oracle compatibility (#1003)
 
 # 3.51.0